+++ parser/typedef.go
@@ -1,11 +1,36 @@
 package parser
 
+type KeyValue struct {
+	node
+	Key   *Name `@@ ( "="`
+	Value *Name `@@ )?`
+}
+
+func (n *KeyValue) Children() []Node {
+	return sliceOf(n.Key, n.Value)
+}
+
+type Annotation struct {
+	node
+	Name      *Name       `@@`
+	KeyValues []*KeyValue `( "("  @@ ("," @@)*  ")" )?`
+}
+
+func (n *Annotation) Children() []Node {
+	res := sliceOf(n.Name)
+	for _, value := range n.KeyValues {
+		res = append(res, value)
+	}
+
+	return res
+}
+
 type TypeDefinition struct {
 	node
 	// Description may be nil
-	Description *Literal  `@@?`
-	Tags        []*Name   `("@" @@)*`
-	Type        NamedType `@@`
+	Description *Literal      `@@?`
+	Annotations []*Annotation `("@" @@)*`
+	Type        NamedType     `@@`
 }
 
 func TypeDefinitionFrom(n Node) *TypeDefinition {
@@ -22,7 +47,7 @@ func TypeDefinitionFrom(n Node) *TypeDefinition {
 
 func (n *TypeDefinition) Children() []Node {
 	res := sliceOf(n.Description)
-	for _, tag := range n.Tags {
+	for _, tag := range n.Annotations {
 		res = append(res, tag)
 	}